requirements.txt
python-telegram-bot==20.7
instaloader==4.11
yt-dlp==2023.12.30
python-dotenv==1.0.1
acrcloud==1.0.2
